[FIX] #409: 기본 옵션 Hmg 데이터 없을 때 태그 제거
